+++ src/SoftTissue.DataContract/ExperimentalAnalysis/AnalyzeAndExtrapolateResults/AnalyzeAndExtrapolateResultsResponse.cs
@@ -3,7 +3,7 @@
 namespace SoftTissue.DataContract.ExperimentalAnalysis.AnalyzeAndExtrapolateResults
 {
     /// <summary>
-    /// It represents the response content for AnalyzeAndPredictResults operations.
+    /// It represents the response content for AnalyzeAndExtrapolateResults operations.
     /// </summary>
     public sealed class AnalyzeAndExtrapolateResultsResponse : OperationResponseBase<AnalyzeAndExtrapolateResultsResponseData> { }
 }