+++ etc/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2021-04-25
+### Fixed
+- Operation AnalyzeAndExtrapolateResults that was writing invalid results in the solution file and moved main parts that analyze the results to a new operation.
+### Added
+- Operation AnalyzeResults.
+
 ## [2.0.1] - 2021-03-31
 ### Fixed
 - Data Contracts to not use JsonConstructor that was crashing when using Swagger.